data pipeline/backend/fastapi/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
joblib>=1.4.0
numpy>=2.4.0
scikit-learn>=1.5.0
xgboost>=2.1.0
pandas>=2.2.0
imbalanced-learn
lightgbm
pymysql
python-dotenv
